+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 - BrPhone class to validate both masked and unmasked Brazilian landline numbers
   — [108](https://github.com/dartoos-dev/formdator/issues/108).
+- BrMobile class to validate both masked and unmasked Brazilian mobile numbers
+  — [107](https://github.com/dartoos-dev/formdator/issues/107).
 
 ## [0.10.0] - 2021-08-10